# pkg/api/v1/docs/api_swagger.yaml
basePath: /v1
consumes:
- application/json
definitions:
...
  models.UserKeyResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      secret:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: admin@loopholelabs.io
    name: API Support
  description: Returns the health and status of the various services that make up
    the API.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://loopholelabs.io/privacy
  title: EndKey API V1
  version: "1.0"
paths:
...
  /userkey:
    get:
      consumes:
      - application/json
      description: Lists all the User Keys
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserKeyResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - userkey
  /userkey/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a User Key
      parameters:
      - description: User Key Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - userkey
    post:
      consumes:
      - application/json
      description: Create a new User Key
      parameters:
      - description: User Key Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserKeyResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - userkey
  /userkey/rotate/{name}:
    post:
      consumes:
      - application/json
      description: Rotates a given User Key
      parameters:
      - description: User Key Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserKeyResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - userkey
produces:
- application/json
schemes:
- https
swagger: "2.0"